segregating function to get list of images that weren't calibrated by RMS
