+++ archive/ukmon_pylib/analysis/gatherDetectionData.py
@@ -11,18 +11,6 @@ import pandas as pd
 from meteortools.ukmondb import getECSVs, getLiveJpgs, getFBfiles, getDetections
 
 
-def getUncalibratedImageList(dtstr=None):
-    datadir=os.getenv('DATADIR', default=os.path.expanduser('~/prod/data'))
-    logfile = os.path.join(datadir, '..', 'logs', 'matchJob.log')
-    flines = open(logfile, 'r').readlines()
-    uncal = [f for f in flines if 'not recalibrated' in f]
-    imglist = [x.strip(' ').split(' ')[1][:-1] for x in uncal]
-    with open(os.path.join(datadir, 'single', 'used', f'uncal_{dtstr}.txt'), 'w') as outf:
-        for li in imglist:
-            outf.write(f'{li}\n')
-    return imglist
-
-
 def getRawData(idlist, outpth):
     for _,row in idlist.iterrows():
         stat =row.ID
