+++ archive/ukmon_pylib/utils/getUncalImages.py
@@ -1,0 +1,16 @@
+# Copyright (C) 2018-2023 Mark McIntyre
+#
+
+import os 
+
+
+def getUncalibratedImageList(dtstr=None):
+    datadir=os.getenv('DATADIR', default=os.path.expanduser('~/prod/data'))
+    logfile = os.path.join(datadir, '..', 'logs', 'matchJob.log')
+    flines = open(logfile, 'r').readlines()
+    uncal = [f for f in flines if 'not recalibrated' in f]
+    imglist = [x.strip(' ').split(' ')[1][:-1] for x in uncal]
+    with open(os.path.join(datadir, 'single', 'used', f'uncal_{dtstr}.txt'), 'w') as outf:
+        for li in imglist:
+            outf.write(f'{li}\n')
+    return imglist
